Refactor profile mappers and update access control
Moved ProfilesBatchFetchMapper to the 'impl' package and updated its import in ProfileController. Added @Mapper annotation to ProfileFetchMapper for Spring integration. Changed organization access control in ProfilesBatchFetchUseCase to require AccessLevel.READ.
