+++ backend/blacknode-software-backend-api/src/main/java/software/blacknode/backend/api/controller/profile/ProfileController.java
@@ -17,8 +17,8 @@ import lombok.RequiredArgsConstructor;
 import me.hinsinger.hinz.common.huid.HUID;
 import software.blacknode.backend.api.controller.annotation.BearerAuth;
 import software.blacknode.backend.api.controller.organization.annotation.OrganizationHeader;
-import software.blacknode.backend.api.controller.profile.mapper.ProfilesBatchFetchMapper;
 import software.blacknode.backend.api.controller.profile.mapper.impl.ProfileFetchMapper;
+import software.blacknode.backend.api.controller.profile.mapper.impl.ProfilesBatchFetchMapper;
 import software.blacknode.backend.api.controller.profile.request.ProfilesBatchFetchRequest;
 import software.blacknode.backend.api.controller.profile.response.ProfileResponse;
 import software.blacknode.backend.api.controller.profile.response.ProfilesBatchFetchResponse;
